==> Step7_DownloadGBIF.Rmd <==
---
title: "Step1_DownloadGBIF"
author: "Josie Lesage"
date: "10/13/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = TRUE)

# Setup packages
library(rgbif)
library(tidyverse)
```

# File goal
This file is intended to gather GBIF records for species on the California Channel Islands, which will be used to update the Cal-IBIS repository. **Before getting started**, create a folder for the date at which you are doing the download (for example, "Dec 2020"). 

The file location should look something like:
> "C:\Users\Josie\Dropbox\SBBG\Cal-IBIS\Cal_IBIS\Methods for Updating Collections\SBBG-Cal-IBIS-GBIF\Data\**Month YY**"

Capitalization matters here, so be careful and consistent. 

Then, within that folder, **create a "Raw" folder and "Clean Data" folder** -- this is where all of your files will end up.

Finally, hit ctrl+F within here, and *replace all* of the previous dates (for instance, "Dec 2020") with the new date ("Mar 2020").

# Gathering GBIF Data
To begin, we'll gather the data we're interested in from GBIF. We'll download several separate datasets - these will end up as .zip folders in the "Data/GBIF/Month YY" folder we created, and will need to be unzipped later.

Run the first chunk below. The download will take a while. Be patient -- give it ~20 minutes before coming back and moving on to the next chunk. Josie suggests making a cup of nice tea and enjoying it while tidying your inbox.

The chunk below is very long, but will prep your data downloads for everything in the following order:
1. Reptiles
2. Mammals
3. Fungi
4. Arthropods
5. Plants
6. Amphibia
7. Aves
8. Fishes
9. Non-arthro inverts

```{r prep all data downloads}
all_dl <- occ_download_queue(

# REPTILIA
occ_download(user = "jlesage", 
                  pwd = "GBIF4SBBG",
                  email = "jlesage@sbbg.org",
                  format = "DWCA",
                  pred("taxonKey", 358),
                  pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                  )),

# MAMMALS
occ_download(user = "jlesage", 
                  pwd = "GBIF4SBBG",
                  email = "jlesage@sbbg.org",
                  format = "DWCA",
                  pred("taxonKey", 359),pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                  )),

# Fungi
occ_download(user = "jlesage", 
                  pwd = "GBIF4SBBG",
                  email = "jlesage@sbbg.org",
                  format = "DWCA",
                  pred("taxonKey", 5), 
                  pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                  )),
              
# Arthropods
occ_download(user = "jlesage", 
                  pwd = "GBIF4SBBG",
                  email = "jlesage@sbbg.org",
                  format = "DWCA",
                  pred("taxonKey", 54), 
                  pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                  )),
                  
                  
# Plants
occ_download(user = "jlesage", 
                  pwd = "GBIF4SBBG",
                  email = "jlesage@sbbg.org",
                  format = "DWCA",
                  pred("taxonKey", 6), 
                  pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                  )),

# Amphibia
occ_download(user = "jlesage", 
                  pwd = "GBIF4SBBG",
                  email = "jlesage@sbbg.org",
                  format = "DWCA",
                  pred("taxonKey", 131),
                  pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                  )),

# AVES
occ_download(user = "jlesage", 
             pwd = "GBIF4SBBG",
             email = "jlesage@sbbg.org",
             format = "DWCA",
             pred("taxonKey", 212),
             pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                         )),

# FISHES
occ_download(user = "jlesage", 
             pwd = "GBIF4SBBG",
             email = "jlesage@sbbg.org",
             format = "DWCA",
             pred("taxonKey", 119), pred("taxonKey", 239), pred("taxonKey", 121), pred("taxonKey", 120), pred("taxonKey", 204), pred("taxonKey", 238), 
             pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                         )),


# Non-insect inverts
occ_download(user = "jlesage", 
             pwd = "GBIF4SBBG",
             email = "jlesage@sbbg.org",
             format = "DWCA",
             pred_or(pred("taxonKey", 105), pred("taxonKey", 108), pred("taxonKey", 110), pred("taxonKey", 14),
                     pred("taxonKey", 19), pred("taxonKey", 22), pred("taxonKey", 42), pred("taxonKey", 43),
                     pred("taxonKey", 45), pred("taxonKey", 50), pred("taxonKey", 51), pred("taxonKey", 52),
                     pred("taxonKey", 53), pred("taxonKey", 55), pred("taxonKey", 5967454), 
                     pred("taxonKey", 5967456), pred("taxonKey", 5967481), pred("taxonKey", 62), 
                     pred("taxonKey", 63), pred("taxonKey", 64), pred("taxonKey", 67), pred("taxonKey", 7190138),
                     pred("taxonKey", 74), pred("taxonKey", 75), pred("taxonKey", 76), pred("taxonKey", 7664204),
                     pred("taxonKey", 77), pred("taxonKey", 8173593), pred("taxonKey", 8246594), 
                     pred("taxonKey", 8355438), pred("taxonKey", 91), pred("taxonKey", 9238047)),
             pred_within("MULTIPOLYGON(((-115.058 27.8699, -115.091 27.8514, -115.090 27.8420, -115.067 27.8200, -115.083 27.7939, -115.108 27.7734, -115.161 27.7587, -115.228 27.7701, -115.300 27.8337, -115.320 27.8642, -115.323 27.9120, -115.299 27.9546, -115.337 27.9832, -115.376 27.9860, -115.413 28.0018, -115.439 28.0244, -115.456 28.0547, -115.459 28.0889, -115.443 28.1547, -115.421 28.1820, -115.387 28.2010, -115.353 28.2411, -115.365 28.3003, -115.338 28.3900, -115.296 28.4386, -115.245 28.4629, -115.190 28.4635, -115.138 28.4388, -115.103 28.4003, -115.075 28.3287, -115.061 28.2673, -115.067 28.2291, -115.054 28.1866, -115.056 28.1297, -115.080 28.0787, -115.085 28.0076, -115.103 27.9768, -115.127 27.9553, -115.079 27.9225, -115.058 27.8699)),
                           ((-118.265 29.2801, -118.206 29.2647, -118.166 29.2223, -118.159 29.1858, -118.169 29.1446, -118.128 29.0784, -118.120 28.9596, -118.134 28.8871, -118.198 28.7949, -118.238 28.7661, -118.289 28.7572, -118.321 28.7631, -118.352 28.7786, -118.393 28.8309, -118.400 28.8608, -118.396 28.9051, -118.415 28.9639, -118.415 28.9944, -118.448 29.0227, -118.465 29.0563, -118.462 29.1126, -118.469 29.1501, -118.451 29.1983, -118.395 29.2438, -118.265 29.2801)),
                           ((-115.695 29.8341, -115.709 29.7550, -115.700 29.7433, -115.744 29.7049, -115.782 29.6935, -115.824 29.6958, -115.860 29.7114, -115.887 29.7376, -115.900 29.7752, -115.895 29.8147, -115.875 29.8498, -115.844 29.8736, -115.802 29.8854, -115.758 29.8815, -115.722 29.8638, -115.695 29.8341)),
                           ((-116.043 30.5660, -116.058 30.4715, -116.040 30.4325, -116.025 30.4225, -116.058 30.4014, -116.103 30.3903, -116.146 30.3954, -116.189 30.4146, -116.216 30.4439, -116.228 30.4810, -116.223 30.5177, -116.202 30.5508, -116.166 30.5756, -116.125 30.5870, -116.081 30.5833, -116.043 30.5660)),
                           ((-116.695 31.7428, -116.724 31.7577, -116.747 31.7568, -116.753 31.7407, -116.726 31.7149, -116.765 31.7011, -116.809 31.7017, -116.850 31.7176, -116.894 31.7546, -116.913 31.7872, -116.914 31.8280, -116.894 31.8657, -116.860 31.8914, -116.817 31.9032, -116.775 31.9001, -116.726 31.8804, -116.699 31.8552, -116.684 31.8219, -116.680 31.7786, -116.695 31.7428)),
                           ((-117.236 32.3001, -117.275 32.3050, -117.308 32.3198, -117.338 32.3509, -117.377 32.3760, -117.400 32.4090, -117.407 32.4477, -117.398 32.4824, -117.373 32.5116, -117.337 32.5308, -117.293 32.5363, -117.251 32.5264, -117.181 32.4892, -117.144 32.4462, -117.133 32.3943, -117.146 32.3476, -117.183 32.3140, -117.236 32.3001)),
                           ((-118.343 32.7261, -118.414 32.7107, -118.462 32.7156, -118.492 32.7299, -118.587 32.7994, -118.711 32.9949, -118.709 33.0590, -118.694 33.0844, -118.670 33.1050, -118.598 33.1258, -118.550 33.1211, -118.515 33.1083, -118.477 33.0774, -118.416 32.9937, -118.263 32.8735, -118.243 32.8295, -118.252 32.7828, -118.291 32.7437, -118.343 32.7261)),
                           ((-119.537 33.3755, -119.491 33.3695, -119.409 33.3361, -119.363 33.3033, -119.340 33.2730, -119.327 33.2353, -119.331 33.1991, -119.356 33.1624, -119.389 33.1400, -119.463 33.1248, -119.532 33.1333, -119.609 33.1599, -119.646 33.1876, -119.669 33.2168, -119.686 33.2800, -119.667 33.3300, -119.620 33.3617, -119.537 33.3755)),
                           ((-118.610 33.5691, -118.501 33.5620, -118.456 33.5367, -118.339 33.4976, -118.289 33.4692, -118.262 33.4269, -118.218 33.3824, -118.199 33.3399, -118.201 33.2852, -118.231 33.2435, -118.277 33.2174, -118.334 33.2090, -118.401 33.2286, -118.488 33.2359, -118.542 33.2631, -118.577 33.3020, -118.596 33.3480, -118.656 33.3815, -118.695 33.4261, -118.712 33.4621, -118.711 33.5009, -118.690 33.5354, -118.654 33.5598, -118.610 33.5691)),
                           ((-119.032 33.3750, -119.077 33.3815, -119.118 33.4043, -119.143 33.4384, -119.156 33.4839, -119.146 33.5211, -119.119 33.5517, -119.084 33.5688, -119.028 33.5792, -118.983 33.5702, -118.947 33.5480, -118.925 33.5151, -118.920 33.4753, -118.931 33.4346, -118.954 33.4045, -118.989 33.3837, -119.032 33.3750)),
                           ((-115.572 28.4047, -115.529 28.4044, -115.491 28.3906, -115.454 28.3574, -115.437 28.3242, -115.437 28.2852, -115.452 28.2495, -115.482 28.2193, -115.522 28.2031, -115.604 28.2068, -115.649 28.2187, -115.685 28.2449, -115.705 28.2818, -115.701 28.3343, -115.674 28.3781, -115.629 28.4020, -115.572 28.4047)),
                           ((-120.236 34.1001, -120.208 34.0996, -120.155 34.1152, -120.114 34.1144, -120.070 34.1271, -120.013 34.1224, -119.977 34.1537, -119.927 34.1675, -119.838 34.1602, -119.795 34.1477, -119.737 34.1473, -119.658 34.1216, -119.612 34.1406, -119.551 34.1453, -119.514 34.1361, -119.453 34.1050, -119.354 34.1072, -119.304 34.0944, -119.261 34.0566, -119.250 34.0053, -119.261 33.9746, -119.284 33.9497, -119.363 33.9154, -119.436 33.9157, -119.485 33.9290, -119.528 33.9098, -119.628 33.8971, -119.713 33.8695, -119.828 33.8699, -119.887 33.8830, -119.920 33.8616, -119.970 33.8516, -120.011 33.8305, -120.087 33.8072, -120.132 33.8048, -120.171 33.8149, -120.248 33.8540, -120.298 33.9173, -120.317 33.9287, -120.373 33.9247, -120.488 33.9477, -120.528 33.9712, -120.554 34.0086, -120.556 34.0522, -120.533 34.0941, -120.474 34.1313, -120.404 34.1616, -120.364 34.1664, -120.325 34.1589, -120.236 34.1001)))"
                         ))
)


```

## Downloading zip files
Did you wait at least 20 minutes?

If you've waited, run the chunk below. It should show 9 datasets, and should show "succeeded" on the third page of results for each (you'll need to use the arrow keys to flip between pages). If it hasn't finished, indicates that things are still running, or that some requests haven't started, you should wait a little, and then check again (by running the chunk below again). 

If you get a failure message, call Josie to troubleshoot.

When the chunk below shows that all data has been prepared successfully, you're ready to move on. This chunk will also show the "keys" are associated with each of the 9 requests - in the reverse order that you made the requests in.

```{r show download list}
occ_download_list(user = "jlesage", 
             pwd = "GBIF4SBBG", limit = 9)
```

You can now download the data! However, we need to make a few changes to the code below:

**BEFORE RUNNING THE CHUNK:** You'll need to change the keys (the long numbers in quotes) to reflect the keys shown by the chunk above - just copy and paste each of those nine below. 

Also, make sure the date in the file path is correct!

Downloading the .zip files by running the code below will also take a small chunk of time -- be patient while the computer obtains all of the info.

```{r}
occ_download_get("0132501-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132502-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132503-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132504-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132505-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132506-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132507-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132547-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
occ_download_get("0132550-200613084148143",
                 overwrite = TRUE,
                 path = "Data/Dec 2020/GBIF/Raw")
```


Once all 9 zip folders are downloaded, you'll need to rename them for the correct group in your computer file browser (meaning this work can't happen in R). It's easiest to do this by comparing with the downloads list online:
https://www.gbif.org/user/download

The DOI url visible online should line up with the occ_download_list chunk above, and you'll be able to see exactly which groups were called on. 

Reminder that the downloads were as follows:
1. Reptiles
2. Mammals
3. Fungi
4. Arthropods
5. Plants
6. Amphibia
7. Aves
8. Fish
9. Inverts

Rename the .zip files in the folders to the correct group it belongs to using the GBIF link that shows the DOI (https://www.gbif.org/user/download), occ_download_list below, and the files in the folder. 

After renaming, extract the zip files. The folder should now be "DATE/Raw" with 9 folders (one for each group, as named above).

Congrats! You've completed step 1 of the process. Time to move on to step 2.
